+++ pkg/deployment/reconcile/plan_builder_tls.go
@@ -572,7 +572,7 @@ func (r *Reconciler) keyfileRenewalRequired(ctx context.Context, apiObject k8sut
 
 	// Ensure secret is propagated only on 3.7.0+ enterprise and inplace mode
 	if mode == api.TLSRotateModeInPlace && group.IsArangod() {
-		conn, err := context.GetServerClient(ctx, group, member.ID)
+		conn, err := context.GetMembersState().GetMemberClient(member.ID)
 		if err != nil {
 			r.planLogger.Err(err).Warn("Unable to get client")
 			return false, false