+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 - (Feature) Member restarts metric
 - (Bugfix) Infinite loop fix in ArangoD AsyncClient
 - (Bugfix) Add Panic Handler
+- (Refactor) Use cached member's clients 
 
 ## [1.2.13](https://github.com/arangodb/kube-arangodb/tree/1.2.13) (2022-06-07)
 - (Bugfix) Fix arangosync members state inspection